+++ src/mflux/utils/version_util.py
@@ -8,7 +8,7 @@ import toml
 
 
 class VersionUtil:
-    PACKAGED_RELEASE_DATE = "2026-08-31"
+    PACKAGED_RELEASE_DATE = "2026-09-01"
     _version: str | None = None
     _release_date: str | None = None